+++ redmine/engines/__init__.py
@@ -1,0 +1,6 @@
+"""
+Defines engines for processing requests/responses to/from Redmine.
+"""
+
+from .base import BaseEngine
+from .sync import SyncEngine