+++ tests/__init__.py
@@ -10,4 +10,22 @@ except ImportError:
 
 from redmine import Redmine
 
-URL = 'http://foo.bar'
+
+class BaseRedmineTestCase(unittest.TestCase):
+    url = 'http://foo.bar'
+    patch_prefix = 'patch'
+    patch_targets = {'requests': 'redmine.engines.sync.requests.Session.request'}
+
+    def setUp(self):
+        self.redmine = Redmine(self.url)
+        self.response = mock.Mock(status_code=200)
+
+        for target, path in self.patch_targets.items():
+            setattr(self, '{0}_{1}'.format(self.patch_prefix, target),
+                    mock.patch(path, return_value=self.response).start())
+
+        self.addCleanup(mock.patch.stopall)
+
+    def set_patch_side_effect(self, side_effect):
+        for target in self.patch_targets:
+            getattr(self, '{0}_{1}'.format(self.patch_prefix, target)).side_effect = side_effect